+++ packages/column-views/src/unit-details/panel.ts
@@ -226,20 +226,7 @@ function UnitDetailsContent({
       h(Parenthetical, h(Duration, { value: unit.b_age - unit.t_age })),
       h(IntervalProportions, { 
         unit,
-        onClick: (e) => {
-          const child = e.target;
-          const parent = e.target.parentElement;;
-          const grandparent = parent.parentElement;
-          const greatgrandparent = grandparent.parentElement;
-
-          const check = [child, parent, grandparent, greatgrandparent].filter(
-            (el) => el.className.includes("int_id")
-          );
-
-          const id = check[0]?.className.match(/int_id-(\d+)/)?.[1];
-
-          window.open("/lex/intervals/" + id, "_blank");
-        },
+        onClickItem,
        }),
     ]),
     h(EnvironmentsList, { 
@@ -455,7 +442,7 @@ function UnitIDList({ units, selectUnit }) {
   );
 }
 
-function IntervalProportions({ unit, onClick }) {
+function IntervalProportions({ unit, onClickItem }) {
   const i0 = unit.b_int_id;
   const i1 = unit.t_int_id;
   let b_prop = unit.b_prop ?? 0;
@@ -483,20 +470,26 @@ function IntervalProportions({ unit, onClick }) {
     p0 = h("span.joint-proportion", [p0, h("span.sep", "to"), p1]);
   }
 
-  const clickable = onClick != null;
+  const clickable = onClickItem != null;
+
+  const handleClick = (event: MouseEvent) => {
+    if (onClickItem) {
+      onClickItem({ event, int_id: interval0.int_id });
+    }
+  };
 
   return h("div.interval-proportions", [
     h(IntervalTag, {
-      className: "int_id-" + interval0?.int_id + (clickable ? " clickable" : ""),
-      onClick,
+      className: clickable ? "clickable" : "",
+      onClick: clickable ? handleClick : undefined,
       interval: interval0,
       prefix: p0,
     }),
     h.if(i0 != i1)("span.discourage-break", [
       h("span.sep", "to"),
       h(IntervalTag, {
-        className: "int_id-" + interval0?.int_id + (clickable ? " clickable" : ""),
-        onClick,
+        className: clickable ? "clickable" : "",
+        onClick: clickable ? handleClick : undefined,
         interval: {
           ...int1,
           id: i1,