+++ packages/data-components/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [unreleased]
 
-- onClick capability added to LithologyTag, where event returns field `lithId` with corresponding lith_id
+- onClick capability added to LithologyTag, where event returns field `lith_id` with corresponding lith_id
+- onClick capability added to IntervalTag, where event returns field `int_id` with corresponding int_id
 
 ## [0.1.0] - 2025-04-09